pyproject.toml
[project]
name = "endtoendlang-chainproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "flask>=3.1.3",
    "google-genai>=1.73.1",
    "hf-xet>=1.4.3",
    "huggingface-hub>=1.12.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.2.1",
    "langchainhub>=0.1.21",
    "langserve>=0.3.3",
    "langsmith>=0.7.37",
    "numpy>=2.4.4",
    "ollama>=0.6.1",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20260107",
    "pi-heif>=1.3.0",
    "pymupdf>=1.27.2.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "setuptools>=82.0.1",
    "sse-starlette>=3.4.1",
    "streamlit>=1.56.0",
    "transformers>=5.6.2",
    "unstructured>=0.22.23",
    "unstructured-inference>=1.6.10",
    "uvicorn>=0.46.0",
    "wikipedia>=1.4.0",
]
